Image Loader + Empty Object Components Ignore
If a component is an empty object it will be ignored when populating data. If an object is empty, the API is returning it because we have stripped it out when determining whether the component should be returned to that user. I little hacky fix.
Image loader style fixes for horizontal images when the 'cover' prop is set to true.